Fixed making snapshots. Moved snapShot method to liveStreamHandler
