+++ src/main/java/nl/tudelft/contextproject/camera/CameraConnection.java
@@ -28,14 +28,6 @@ public abstract class CameraConnection implements Observer {
      * @return The link to a live feed from the camera.
      */
     public abstract String getStreamLink();
-    
-    /**
-     * Takes a snapshot of the image present on a camera and stores the image
-     * at a chosen location.
-     *
-     * @param imageLocation The location to which the captured image is stored.
-     */
-    public abstract void snapShot(String imageLocation);
 
     /**
      * Returns the current camera settings of the camera as a
